+++ odsp-shared-react/src/components/CardList/renderers/DocumentCardTile.tsx
@@ -44,7 +44,8 @@ export class DocumentCardTile extends React.Component<ICardTileProps, {}> {
       onClick,
       onClickHref,
       customIconAcronym,
-      customIconBgColor
+      customIconBgColor,
+      className
     } = this.props.item;
 
     const showPreview = !customIconAcronym || !customIconBgColor;
@@ -57,23 +58,13 @@ export class DocumentCardTile extends React.Component<ICardTileProps, {}> {
     }
 
     return (
-      <DocumentCard onClick={ onClick } onClickHref={ onClickHref } type={ DocumentCardType.compact } accentColor={ accentColor } >
+      <DocumentCard onClick={ onClick } onClickHref={ onClickHref } type={ DocumentCardType.compact } accentColor={ accentColor } className={ className }>
         { showPreview &&
           <DocumentCardPreview
             previewImages = { previewImages }
             getOverflowDocumentCountText={ getOverflowDocumentCountText }/>
         }
-        { !showPreview &&
-          <div className='ms-DocumentCardPreview' style={ { 'borderBottomColor': customIconBgColor } }>
-            <div
-              role='presentation'
-              aria-hidden='true'
-              className='ms-DocumentCardTile-customIcon ms-font-xxl'
-              style={ { 'backgroundColor': customIconBgColor } } >
-              { customIconAcronym }
-            </div>
-          </div>
-        }
+        { !showPreview && this._renderCustomIconPreview(customIconBgColor, customIconAcronym) }
         <div className='ms-DocumentCard-details'>
           <DocumentCardTitle title={ title } shouldTruncate={ true }/>
           { people && people.length > 0 &&
@@ -101,29 +92,20 @@ export class DocumentCardTile extends React.Component<ICardTileProps, {}> {
       onClick,
       onClickHref,
       customIconAcronym,
-      customIconBgColor
+      customIconBgColor,
+      className
     } = this.props.item;
 
     const showPreview = !customIconAcronym || !customIconBgColor;
 
     return (
-      <DocumentCard onClick={ onClick } onClickHref={ onClickHref } >
+      <DocumentCard onClick={ onClick } onClickHref={ onClickHref } className={ className }>
         { showPreview &&
           <DocumentCardPreview
             previewImages = { previewImages }
             getOverflowDocumentCountText={ getOverflowDocumentCountText }/>
         }
-        { !showPreview &&
-          <div className='ms-DocumentCardPreview' style={ { 'borderBottomColor': customIconBgColor } }>
-            <div
-              role='presentation'
-              aria-hidden='true'
-              className='ms-DocumentCardTile-customIcon ms-font-xxl'
-              style={ { 'backgroundColor': customIconBgColor } } >
-              { customIconAcronym }
-            </div>
-          </div>
-        }
+        { !showPreview && this._renderCustomIconPreview(customIconBgColor, customIconAcronym) }
         <div className={ css({ 'has-location': !!location && !hideLocation }, 'ms-DocumentCardTile-titleArea') }>
           { location && !hideLocation &&
             <DocumentCardLocation location={ location } onClick={ locationOnClick } locationHref={ locationHref }/>
@@ -157,5 +139,19 @@ export class DocumentCardTile extends React.Component<ICardTileProps, {}> {
          window.location.href = onClickHref;
       }
     }
+  }
+
+  private _renderCustomIconPreview(customIconBgColor: string, customIconAcronym: string) {
+    return (
+      <div className='ms-DocumentCardPreview ms-DocumentCardTile-customIconPreview' style={ { 'borderBottomColor': customIconBgColor } }>
+        <div
+          role='presentation'
+          aria-hidden='true'
+          className='ms-DocumentCardTile-customIcon'
+          style={ { 'backgroundColor': customIconBgColor } } >
+          { customIconAcronym }
+        </div>
+      </div>
+    )
   }
 }