+++ odsp-shared-react/src/containers/filterSelectPanel/index.ts
@@ -1,0 +1,2 @@
+export * from './FilterSelectPanel';
+export * from './FilterSelectPanel.Props';